# src/main/java/com/exercise/inventories/api/InventoriesApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.18).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.exercise.inventories.api;

import com.exercise.inventories.model.Inventories;
import com.exercise.inventories.model.Inventory;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.bind.annotation.CookieValue;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
@Api(value = "inventories", description = "the inventories API")
public interface InventoriesApi {

    @ApiOperation(value = "Add a new inventory", nickname = "addInventory", notes = "", response = Inventory.class, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "The inventory added", response = Inventory.class) })
    @RequestMapping(value = "/inventories",
        produces = { "*/*", "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.PUT)
    ResponseEntity<Inventory> addInventory(@ApiParam(value = "" ,required=true )  @Valid @RequestBody Inventory body
);


    @ApiOperation(value = "Get all inventories", nickname = "getAllInventories", notes = "", response = Inventories.class, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "All inventories", response = Inventories.class) })
    @RequestMapping(value = "/inventories",
        produces = { "*/*", "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Inventories> getAllInventories();


    @ApiOperation(value = "Get an inventory by inventoryId", nickname = "getInventoryById", notes = "", response = Inventory.class, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Get inventory with the given ID", response = Inventory.class) })
    @RequestMapping(value = "/inventories/{inventoryId}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Inventory> getInventoryById(@ApiParam(value = "inventory ID",required=true) @PathVariable("inventoryId") Integer inventoryId
);

}