# src/main/java/com/exercise/inventories/model/Inventories.java
package com.exercise.inventories.model;

import java.util.Objects;
import com.exercise.inventories.model.Inventory;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import java.util.ArrayList;
import java.util.List;
import org.springframework.validation.annotation.Validated;
import javax.validation.Valid;
import javax.validation.constraints.*;

/**
 * Inventories
 */
@Validated
public class Inventories   {
  @JsonProperty("data")
  @Valid
  private List<Inventory> data = null;

  public Inventories data(List<Inventory> data) {
    this.data = data;
    return this;
  }

  public Inventories addDataItem(Inventory dataItem) {
    if (this.data == null) {
      this.data = new ArrayList<Inventory>();
    }
    this.data.add(dataItem);
    return this;
  }

  /**
   * Get data
   * @return data
  **/
  @ApiModelProperty(value = "")
      @Valid
    public List<Inventory> getData() {
    return data;
  }

  public void setData(List<Inventory> data) {
    this.data = data;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    Inventories inventories = (Inventories) o;
    return Objects.equals(this.data, inventories.data);
  }

  @Override
  public int hashCode() {
    return Objects.hash(data);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class Inventories {\n");
    
    sb.append("    data: ").append(toIndentedString(data)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}